+++ packages/hummingbird/src/cards/CollectionCard.js
@@ -1,1 +1,0 @@
-export { CollectionCard as default } from '@wingscms/hummingbird';