+++ packages/hummingbird/src/components/Campaign.js
@@ -1,0 +1,1 @@
+export { Campaign as default } from '@wingscms/react';