+++ packages/react/src/cards/QuoteCard/Quote.js
@@ -1,32 +1,58 @@
 import styled from 'styled-components';
+import { ALIGNLEFT, ALIGNRIGHT } from '@wingscms/crane';
 import { FLOAT } from './enums';
-import { ALIGNLEFT, ALIGNRIGHT } from './floats';
 
-import { quote } from './icons';
+import { quote } from '../../img/icons';
 
 export default styled.figure`
   clear: both;
   width: 100%;
   text-align: center;
   margin: 80px 0;
-  box-shadow: 0 0 40px 0 rgba(0, 0, 0, 0.05);
+  box-shadow: ${({ theme }) => theme.defaultShadow};
   padding: 40px 0 20px 0;
+  background-color: ${({ theme }) => theme.blockquoteBackgroundColor};
+  a { 
+      color: #000000;
+      text-decoration: none;
+      background-image: linear-gradient(
+        120deg,
+        ${({ theme }) => theme.primaryColor} 0%,
+        ${({ theme }) => theme.primaryColor} 100%
+      );
+      background-repeat: no-repeat;
+      background-size: 100% 2px;
+      background-position: 0% 100%;
+      transition: background-size 0.1s linear;
+      &:hover,
+      &:focus {
+        background-size: 100% 4px;
+        background-image: linear-gradient(120deg, ${({ theme }) => theme.primaryColor} 0%, ${({
+  theme,
+}) => theme.primaryColor} 100%);
+      }
+    }
+  @media screen and (max-width: 800px) {
+      margin-top: 40px;
+      margin-bottom: 40px;
+  }
   &.pullquote-${FLOAT.NONE},
   &.pullquote-${FLOAT.LEFT},
   &.pullquote-${FLOAT.RIGHT} {
+    background-color: ${({ theme }) => theme.pullquoteBackgroundColor};
     margin: 40px 0;
+    padding: 20px;
     box-shadow: none;
     aside {
-      font-family: initial;
-      font-size: 40px;
+      font-family: ${({ theme }) => theme.headerFontFamily};
+      font-size: 24px;
       line-height: 1.2;
-      @media screen and (max-width: 800px) {
-        font-size: 24px;
+      @media screen and (min-width: 800px) {
+        font-size: 28px;
       }
     }
   }
   &.pullquote-${FLOAT.NONE} {
-    padding: 0;
     figcaption {
       margin-top: 20px;
     }
@@ -40,29 +66,22 @@ export default styled.figure`
   &.pullquote-${FLOAT.LEFT},
   &.pullquote-${FLOAT.RIGHT} {
     clear: none;
-    padding: 0.75rem;
     text-align: left;
     @media screen and (max-width: 800px) {
       width: 100%;
       float: none;
     }
-    @media screen and (min-width: 400px) {
-      padding: 1.5rem;
-    }
-    @media screen and (min-width: 1080px) {
-      padding: 2rem;
-    }
   }
-  aside,
-  blockquote {
-    color: ${({ theme }) => theme.primaryColor};
+  aside {
+    color: ${({ theme }) => theme.pullquoteColor};
     padding: 0;
-    font-weight: 900;
+    font-weight: ${({ theme }) => theme.pullquoteFontWeight};
   }
   blockquote {
-    color: ${({ theme }) => theme.textColor};
+    color: ${({ theme }) => theme.blockquoteColor};
+    background-color: ${({ theme }) => theme.blockquoteBackgroundColor};
     padding: 0;
-    font-weight: 300;
+    font-weight: ${({ theme }) => theme.blockquoteFontWeight};
     text-align: left;
     position: relative;
     &:after {
@@ -83,8 +102,5 @@ export default styled.figure`
     &:before {
       content: '— ';
     }
-  }
-  a {
-    color: black;
   }
 `;