+++ packages/react/src/cards/QuoteCard/QuoteCard.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import { ThemeProvider } from 'styled-components';
 import createCard from '../../createCard';
 import propTypes from './propTypes';
 import BlockquoteView from './BlockquoteView';
@@ -22,11 +21,7 @@ class QuoteCardView extends Component {
 
   render() {
     const Comp = this.getComponent();
-    return (
-      <ThemeProvider theme={{}}>
-        <Comp {...this.props} />
-      </ThemeProvider>
-    );
+    return <Comp {...this.props} />;
   }
 }