+++ packages/crane/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.5](https://github.com/wingscms/wings/compare/@wingscms/crane@0.14.4...@wingscms/crane@0.14.5) (2019-08-06)
+
+
+### Bug Fixes
+
+* improve Q&A styling ([#238](https://github.com/wingscms/wings/issues/238)) ([97b867e](https://github.com/wingscms/wings/commit/97b867e))
+
+
+
+
+
+## [0.14.4](https://github.com/wingscms/wings/compare/@wingscms/crane@0.14.3...@wingscms/crane@0.14.4) (2019-08-01)
+
+
+### Bug Fixes
+
+* add arrow icon to form select field ([#236](https://github.com/wingscms/wings/issues/236)) ([73c8373](https://github.com/wingscms/wings/commit/73c8373))
+
+
+
+
+
+## [0.14.3](https://github.com/wingscms/wings/compare/@wingscms/crane@0.14.2...@wingscms/crane@0.14.3) (2019-07-24)
+
+
+### Bug Fixes
+
+* update social media icons and use correct email svg ([#228](https://github.com/wingscms/wings/issues/228)) ([785dc1e](https://github.com/wingscms/wings/commit/785dc1e))
+
+
+
+
+
+<a name="0.14.2"></a>
+## [0.14.2](https://github.com/wingscms/wings/compare/@wingscms/crane@0.14.1...@wingscms/crane@0.14.2) (2019-07-11)
+
+
+
+
+**Note:** Version bump only for package @wingscms/crane
+
+<a name="0.14.0"></a>
+# [0.14.0](https://github.com/wingscms/wings/compare/@wingscms/crane@0.13.0...@wingscms/crane@0.14.0) (2019-07-11)
+
+
+### Features
+
+* **hummingbird/crane/react:** big refactor and introduction of new APIs ([#212](https://github.com/wingscms/wings/issues/212)) ([4fce054](https://github.com/wingscms/wings/commit/4fce054))
+
+
+
+
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.13.0](https://github.com/wingscms/wings/compare/@wingscms/crane@0.12.0...@wingscms/crane@0.13.0) (2019-07-03)